+++ packages/flutter_chat_firebase/lib/service/firebase_chat_detail_service.dart
@@ -238,6 +238,9 @@ class FirebaseChatDetailService
       onCancel: () async {
         await _subscription?.cancel();
         _subscription = null;
+        _cumulativeMessages = [];
+        lastChat = chatId;
+        lastMessage = null;
         debugPrint('Canceling messages stream');
       },
     );
@@ -259,14 +262,16 @@ class FirebaseChatDetailService
   /// [pageSize]: The number of messages to fetch.
   /// [chatId]: The ID of the chat.
   @override
-  Future<void> fetchMoreMessage(int pageSize, String chatId) async {
-    if (lastChat == null) {
-      lastChat = chatId;
-    } else if (lastChat != chatId) {
+  Future<void> fetchMoreMessage(
+    int pageSize,
+    String chatId,
+  ) async {
+    if (lastChat != chatId) {
       _cumulativeMessages = [];
       lastChat = chatId;
       lastMessage = null;
     }
+
     // get the x amount of last messages from the oldest message that is in
     // cumulative messages and add that to the list
     var messages = <ChatMessageModel>[];