+++ packages/flutter_chat_view/lib/src/components/chat_image.dart
@@ -35,6 +35,7 @@ class ChatImage extends StatelessWidget {
         height: size,
         child: image.isNotEmpty
             ? CachedNetworkImage(
+                fadeInDuration: Duration.zero,
                 imageUrl: image,
                 fit: BoxFit.cover,
               )