+++ packages/flutter_chat_view/lib/src/config/chat_options.dart
@@ -6,6 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_chat_view/flutter_chat_view.dart';
 import 'package:flutter_chat_view/src/components/chat_image.dart';
 import 'package:flutter_image_picker/flutter_image_picker.dart';
+import 'package:flutter_profile/flutter_profile.dart';
 
 class ChatOptions {
   const ChatOptions({
@@ -50,13 +51,17 @@ Widget _createNewChatButton(
   ChatTranslations translations,
 ) =>
     Padding(
-      padding: const EdgeInsets.all(16.0),
+      padding: const EdgeInsets.all(24.0),
       child: ElevatedButton(
         style: ElevatedButton.styleFrom(
+          backgroundColor: Colors.black,
           minimumSize: const Size.fromHeight(50),
         ),
         onPressed: onPressed,
-        child: Text(translations.newChatButton),
+        child: Text(
+          translations.newChatButton,
+          style: const TextStyle(color: Colors.white),
+        ),
       ),
     );
 
@@ -66,9 +71,38 @@ Widget _createMessageInput(
   ChatTranslations translations,
 ) =>
     TextField(
+      textCapitalization: TextCapitalization.sentences,
       controller: textEditingController,
       decoration: InputDecoration(
+        enabledBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(26.5),
+          borderSide: const BorderSide(
+            color: Colors.black,
+          ),
+        ),
+        focusedBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(26.5),
+          borderSide: const BorderSide(
+            color: Colors.black,
+          ),
+        ),
+        contentPadding: const EdgeInsets.symmetric(
+          vertical: 0,
+          horizontal: 30,
+        ),
         hintText: translations.messagePlaceholder,
+        hintStyle: const TextStyle(
+          fontWeight: FontWeight.normal,
+          color: Colors.black,
+        ),
+        fillColor: Colors.white,
+        filled: true,
+        border: const OutlineInputBorder(
+          borderRadius: BorderRadius.all(
+            Radius.circular(26.5),
+          ),
+          borderSide: BorderSide.none,
+        ),
         suffixIcon: suffixIcon,
       ),
     );
@@ -93,9 +127,13 @@ Widget _createImagePickerContainer(
       color: Colors.white,
       child: ImagePicker(
         customButton: ElevatedButton(
+          style: ElevatedButton.styleFrom(
+            backgroundColor: Colors.black,
+          ),
           onPressed: onClose,
           child: Text(
             translations.cancelImagePickerBtn,
+            style: const TextStyle(color: Colors.white),
           ),
         ),
       ),
@@ -114,8 +152,12 @@ Widget _createUserAvatar(
   ChatUserModel user,
   double size,
 ) =>
-    ChatImage(
-      image: user.imageUrl ?? '',
+    Avatar(
+      user: User(
+        firstName: user.firstName,
+        lastName: user.lastName,
+        imageUrl: user.imageUrl,
+      ),
       size: size,
     );
 Widget _createGroupAvatar(