+++ packages/flutter_chat_view/lib/src/screens/new_chat_screen.dart
@@ -10,6 +10,7 @@ class NewChatScreen extends StatefulWidget {
     required this.options,
     required this.onPressCreateChat,
     required this.service,
+    required this.onPressCreateGroupChat,
     this.translations = const ChatTranslations(),
     super.key,
   });
@@ -22,6 +23,7 @@ class NewChatScreen extends StatefulWidget {
 
   /// Callback function for creating a new chat with a user.
   final Function(ChatUserModel) onPressCreateChat;
+  final Function() onPressCreateGroupChat;
 
   /// Translations for the chat.
   final ChatTranslations translations;
@@ -36,62 +38,148 @@ class _NewChatScreenState extends State<NewChatScreen> {
   String query = '';
 
   @override
-  Widget build(BuildContext context) => Scaffold(
-        appBar: AppBar(
-          title: _buildSearchField(),
-          actions: [
-            _buildSearchIcon(),
-          ],
-        ),
-        body: FutureBuilder<List<ChatUserModel>>(
-          // ignore: discarded_futures
-          future: widget.service.chatUserService.getAllUsers(),
-          builder: (context, snapshot) {
-            if (snapshot.connectionState == ConnectionState.waiting) {
-              return const Center(child: CircularProgressIndicator());
-            } else if (snapshot.hasError) {
-              return Text('Error: ${snapshot.error}');
-            } else if (snapshot.hasData) {
-              return _buildUserList(snapshot.data!);
-            } else {
-              return widget.options
-                  .noChatsPlaceholderBuilder(widget.translations);
-            }
-          },
-        ),
-      );
-
-  Widget _buildSearchField() => _isSearching
-      ? Padding(
-          padding: const EdgeInsets.symmetric(vertical: 8.0),
-          child: TextField(
-            focusNode: _textFieldFocusNode,
-            onChanged: (value) {
-              setState(() {
-                query = value;
-              });
+  Widget build(BuildContext context) {
+    var theme = Theme.of(context);
+    return Scaffold(
+      appBar: AppBar(
+        iconTheme: theme.appBarTheme.iconTheme ??
+            const IconThemeData(color: Colors.white),
+        backgroundColor: theme.appBarTheme.backgroundColor ?? Colors.black,
+        title: _buildSearchField(),
+        actions: [
+          _buildSearchIcon(),
+        ],
+      ),
+      body: Column(
+        children: [
+          GestureDetector(
+            onTap: () async {
+              await widget.onPressCreateGroupChat();
             },
-            decoration: InputDecoration(
-              hintText: widget.translations.searchPlaceholder,
+            child: Container(
+              color: Colors.grey[900],
+              child: SizedBox(
+                height: 60.0,
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.start,
+                      children: [
+                        Padding(
+                          padding: const EdgeInsets.only(
+                            left: 16.0,
+                          ),
+                          child: IconButton(
+                            icon: const Icon(
+                              Icons.group,
+                              color: Colors.white,
+                            ),
+                            onPressed: () {
+                              // Handle group chat creation
+                            },
+                          ),
+                        ),
+                        const Text(
+                          'Create group chat',
+                          style: TextStyle(
+                            color: Colors.white,
+                            fontSize: 16.0,
+                          ),
+                        ),
+                      ],
+                    ),
+                  ],
+                ),
+              ),
             ),
           ),
-        )
-      : Text(widget.translations.newChatButton);
+          Expanded(
+            child: FutureBuilder<List<ChatUserModel>>(
+              // ignore: discarded_futures
+              future: widget.service.chatUserService.getAllUsers(),
+              builder: (context, snapshot) {
+                if (snapshot.connectionState == ConnectionState.waiting) {
+                  return const Center(child: CircularProgressIndicator());
+                } else if (snapshot.hasError) {
+                  return Text('Error: ${snapshot.error}');
+                } else if (snapshot.hasData) {
+                  return _buildUserList(snapshot.data!);
+                } else {
+                  return widget.options
+                      .noChatsPlaceholderBuilder(widget.translations);
+                }
+              },
+            ),
+          ),
+        ],
+      ),
+    );
+  }
 
-  Widget _buildSearchIcon() => IconButton(
-        onPressed: () {
-          setState(() {
-            _isSearching = !_isSearching;
-          });
+  Widget _buildSearchField() {
+    var theme = Theme.of(context);
 
-          if (_isSearching) {
-            _textFieldFocusNode.requestFocus();
-          }
-        },
-        icon: Icon(
-          _isSearching ? Icons.close : Icons.search,
-        ),
-      );
+    return _isSearching
+        ? Padding(
+            padding: const EdgeInsets.symmetric(vertical: 8.0),
+            child: TextField(
+              focusNode: _textFieldFocusNode,
+              onChanged: (value) {
+                setState(() {
+                  query = value;
+                });
+              },
+              decoration: InputDecoration(
+                hintText: widget.translations.searchPlaceholder,
+                hintStyle: theme.inputDecorationTheme.hintStyle ??
+                    const TextStyle(
+                      color: Colors.white,
+                    ),
+                focusedBorder: UnderlineInputBorder(
+                  borderSide: BorderSide(
+                    color: theme.inputDecorationTheme.focusedBorder?.borderSide
+                            .color ??
+                        Colors.white,
+                  ),
+                ),
+              ),
+              style: theme.inputDecorationTheme.hintStyle ??
+                  const TextStyle(
+                    color: Colors.white,
+                  ),
+              cursorColor: theme.textSelectionTheme.cursorColor ?? Colors.white,
+            ),
+          )
+        : Text(
+            widget.translations.newChatButton,
+            style: theme.appBarTheme.titleTextStyle ??
+                const TextStyle(
+                  color: Colors.white,
+                ),
+          );
+  }
+
+  Widget _buildSearchIcon() {
+    var theme = Theme.of(context);
+
+    return IconButton(
+      onPressed: () {
+        setState(() {
+          _isSearching = !_isSearching;
+          query = '';
+        });
+
+        if (_isSearching) {
+          _textFieldFocusNode.requestFocus();
+        }
+      },
+      icon: Icon(
+        _isSearching ? Icons.close : Icons.search,
+        color: theme.appBarTheme.iconTheme?.color ?? Colors.white,
+      ),
+    );
+  }
 
   Widget _buildUserList(List<ChatUserModel> users) {
     var filteredUsers = users
@@ -114,12 +202,15 @@ class _NewChatScreenState extends State<NewChatScreen> {
         var user = filteredUsers[index];
         return GestureDetector(
           child: widget.options.chatRowContainerBuilder(
-            ChatRow(
-              avatar: widget.options.userAvatarBuilder(
-                user,
-                40.0,
+            Container(
+              color: Colors.transparent,
+              child: ChatRow(
+                avatar: widget.options.userAvatarBuilder(
+                  user,
+                  40.0,
+                ),
+                title: user.fullName ?? widget.translations.anonymousUser,
               ),
-              title: user.fullName ?? widget.translations.anonymousUser,
             ),
           ),
           onTap: () async {